move call_service to a util file
